+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 2021-11-30
+
+- Implement script for parsing CLDR gregorian data.
+
 ### 2021-11-29
 
 - Implement script for generating language maps and ordering languages by their popularity.